Fix auth bug

+++ src/comments/comments.service.ts
@@ -6,6 +6,7 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Posts } from 'src/posts/posts.entity';
+import { User } from 'src/users/users.entity';
 import { Like, Repository } from 'typeorm';
 import { Comments } from './comments.entity';
 import { CreateCommentDto } from './dto/comments.dto';
@@ -18,6 +19,8 @@ export class CommentsService {
     private readonly commentsRepository: Repository<Comments>,
     @InjectRepository(Posts)
     private readonly postsRepository: Repository<Posts>,
+    @InjectRepository(User)
+    private readonly usersRepository: Repository<User>,
   ) {}
 
   async findAll(postId: number, queryString: string, order: string) {
@@ -38,7 +41,7 @@ export class CommentsService {
     return comment;
   }
 
-  async create(createDTO: CreateCommentDto) {
+  async create(createDTO: CreateCommentDto, userId: number) {
     const post = await this.postsRepository.findOneBy({ id: createDTO.postId });
 
     if (!post) {
@@ -48,20 +51,37 @@ export class CommentsService {
       );
     }
 
+    const user = await this.postsRepository.findOneBy({ id: userId });
+
+    if (!user) {
+      throw new HttpException(`Invalid user id`, HttpStatus.UNAUTHORIZED);
+    }
+
     const comment = this.commentsRepository.create({
       post: { id: post.id },
+      author: { id: user.id },
       text: createDTO.text,
     });
 
     return await this.commentsRepository.save(comment);
   }
 
-  async update(id: number, updateDTO: UpdateCommentDTO) {
-    const comment = await this.commentsRepository.findOneBy({ id: id });
+  async update(id: number, updateDTO: UpdateCommentDTO, userId: number) {
+    const comment = await this.commentsRepository.findOne({
+      where: { id: id },
+      relations: ['author'],
+    });
     if (!comment) {
       throw new HttpException(
         `Comment with given id = ${id} not found!`,
         HttpStatus.NOT_FOUND,
+      );
+    }
+
+    if (comment.author.id != userId) {
+      throw new HttpException(
+        `Only authors can update comments`,
+        HttpStatus.FORBIDDEN,
       );
     }
 
@@ -72,7 +92,24 @@ export class CommentsService {
     });
   }
 
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
+    const comment = await this.commentsRepository.findOne({
+      where: { id: id },
+      relations: ['author'],
+    });
+    if (!comment) {
+      throw new HttpException(
+        `Comment with given id = ${id} not found!`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    if (comment.author.id != userId) {
+      throw new HttpException(
+        `Only authors can update comments`,
+        HttpStatus.FORBIDDEN,
+      );
+    }
     await this.commentsRepository.delete({ id: id });
   }
 }
